+++ src/components/Buttons/MessagesButton/index.tsx
@@ -1,3 +1,0 @@
-import MessagesButton from './MessagesButton';
-
-export { MessagesButton as default,  };