+++ src/components/Buttons/UpdatesButton/index.tsx
@@ -1,3 +1,0 @@
-import UpdatesButton from './UpdatesButton';
-
-export { UpdatesButton as default,  };