+++ src/hooks/index.ts
@@ -6,5 +6,7 @@ export * from './layout';
 export * from './useAndroidBackButton';
 export * from './useFilter';
 export * from './useFilters';
+export * from './useMenuKeyboardNavigation';
+export * from './useRipple';
 export * from './useSort';
 export * from './useDataTableState';