+++ src/components/Buttons/CodeButton/CodeButton.tsx
@@ -27,13 +27,14 @@ const CodeButton: FC<ButtonProps> = ({ ...restOfProps }) => {
           }}
           color="secondary"
           className="app-code"
+          data-testid="current-instance-code"
           aria-label={t('ui.accessibility.copyInstanceCode', { code })}
           {...restOfProps}
         >
           {`${t('instance.chip')}: ${code}`}
           <AppIcon icon="copy" size="small" sx={{ ml: 1 }} />
         </Button>
-        <Divider role="presentation" />
+        <div className="border-t border-theme-grey mx-3" />
       </>
     )
   );