+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - **Technical improvements**
   - Make tests work in a multi-instance environment (specify instance with envvar INSTANCE_CODE)
+  - Better breadcrumb navigation for mobile
 
 ## 1.8.1